[test] Move/unify pl-l empty geometry tests functions
