+++ test/algorithms/distance/distance_pointlike_linear.cpp
@@ -17,7 +17,7 @@
 #include <boost/test/included/unit_test.hpp>
 
 #include "test_distance_common.hpp"
-
+#include "test_empty_geometry.hpp"
 
 typedef bg::model::point<double,2,bg::cs::cartesian>  point_type;
 typedef bg::model::multi_point<point_type>            multi_point_type;
@@ -268,40 +268,6 @@ void test_distance_multipoint_segment(Strategy const& strategy)
                   "segment(0 0,1 1)",
                   sqrt(2.0), 2, strategy);
 }
-
-//===========================================================================
-
-template <typename Point, typename Strategy>
-void test_more_empty_input_pointlike_linear(Strategy const& strategy)
-{
-#ifdef BOOST_GEOMETRY_TEST_DEBUG
-    std::cout << std::endl;
-    std::cout << "testing on empty inputs... " << std::flush;
-#endif
-    bg::model::linestring<Point> line_empty;
-    bg::model::multi_point<Point> multipoint_empty;
-    bg::model::multi_linestring<bg::model::linestring<Point> > multiline_empty;
-
-    Point point = from_wkt<Point>("point(0 0)");
-    bg::model::linestring<Point> line =
-        from_wkt<bg::model::linestring<Point> >("linestring(0 0,1 1)");
-
-    // 1st geometry is empty
-    test_empty_input(multipoint_empty, line, strategy);
-
-    // 2nd geometry is empty
-    test_empty_input(point, line_empty, strategy);
-    test_empty_input(point, multiline_empty, strategy);
-
-    // both geometries are empty
-    test_empty_input(multipoint_empty, line_empty, strategy);
-    test_empty_input(multipoint_empty, multiline_empty, strategy);
-
-#ifdef BOOST_GEOMETRY_TEST_DEBUG
-    std::cout << "done!" << std::endl;
-#endif
-}
-
 
 //===========================================================================
 //===========================================================================
