+++ test/algorithms/distance/distance_se_pl_l.cpp
@@ -19,7 +19,7 @@
 #include <boost/test/included/unit_test.hpp>
 
 #include "test_distance_se_common.hpp"
-
+#include "test_empty_geometry.hpp"
 
 typedef bg::cs::spherical_equatorial<bg::degree> cs_type;
 typedef bg::model::point<double, 2, cs_type> point_type;
@@ -478,40 +478,6 @@ void test_distance_multipoint_segment(Strategy const& strategy)
                                          strategy),
                   strategy);
 }
-
-//===========================================================================
-
-template <typename Point, typename Strategy>
-void test_more_empty_input_pointlike_linear(Strategy const& strategy)
-{
-#ifdef BOOST_GEOMETRY_TEST_DEBUG
-    std::cout << std::endl;
-    std::cout << "testing on empty inputs... " << std::flush;
-#endif
-    bg::model::linestring<Point> line_empty;
-    bg::model::multi_point<Point> multipoint_empty;
-    bg::model::multi_linestring<bg::model::linestring<Point> > multiline_empty;
-
-    Point point = from_wkt<Point>("POINT(0 0)");
-    bg::model::linestring<Point> line =
-        from_wkt<bg::model::linestring<Point> >("LINESTRING(0 0,1 1)");
-
-    // 1st geometry is empty
-    test_empty_input(multipoint_empty, line, strategy);
-
-    // 2nd geometry is empty
-    test_empty_input(point, line_empty, strategy);
-    test_empty_input(point, multiline_empty, strategy);
-
-    // both geometries are empty
-    test_empty_input(multipoint_empty, line_empty, strategy);
-    test_empty_input(multipoint_empty, multiline_empty, strategy);
-
-#ifdef BOOST_GEOMETRY_TEST_DEBUG
-    std::cout << "done!" << std::endl;
-#endif
-}
-
 
 //===========================================================================
 //===========================================================================
